yarr. fixed a bug where regularizer was incorrectly computed for piecewise exponential due to division of two ints thing in C.

+++ src/piecewise_exponential_rate_function.cpp
@@ -126,7 +126,7 @@ void PiecewiseExponentialRateFunction<T>::compute_regularizer()
     {
         for (int i = 1; i < delta + 1; ++i)
         {
-            xx = (i / delta) * (ts[k + 1] - ts[k]) + ts[k];
+            xx = ((double)i / delta) * (ts[k + 1] - ts[k]) + ts[k];
             etax = 1. / eta(xx);
             tmp = etax - elast;
             _reg += tmp * tmp * exp(-R(xx));
